62112 - Criação que classe de herança para mensagens
